Refactor extend() in es6 to use for-of loop
Since we are getting practice using ES6 features, this code block uses
rest parameters (instead of the arguments object) and uses a for-of loop
to iterate through the array produced by Object.keys(). Since this array
is iterable, we avoid the prior problem of object keys being uniterable
and thus not allowed as a target for the for-of loop.